fix(config): generate
this commit ensures that an existing nginx.conf is not overwritten by
the config generator of the buildpack. for cases where a consumer of
the buildpack used an empty nginx.conf to trigger inclusion of the
buildpack in the plan, the generator will generate a full nginx.conf
according to the buildpack's defaults.

related: #667