Fix bug in retrieve EOL handling
The EOL handling is not working for Nginx. The API call we're making is returning "false" for versions that are not presently EOL. I suspect this is because Nginx does not have a specific EOL date that is announced. It instead will generally cut a new release around April of the year. I changed the API call to one that will return the EOL date if it is present and the release date. Using this, we can calculate a fallback date of the April 30th, for the year after the release. This should be roughly correct, and then when the API has an actual date, we will use that instead.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>